feat: update to v2 of GDF; set homepage to be dynamic based on the first page
